feat(showList): Create ShowsListView class
