+++ AmazingApp/AmazingApp/ShowsList/ShowsListViewControllerViewController.swift
@@ -8,115 +8,11 @@ import UIKit
 
 final class ShowsListViewController: UIViewController {
     
-    private let activityIndicator: UIActivityIndicatorView = {
-        let activityIndicator = UIActivityIndicatorView()
-        
-        activityIndicator.style = .medium
-        activityIndicator.translatesAutoresizingMaskIntoConstraints = false
-        
-        return activityIndicator
-    }()
-    
-    private lazy var refreshControl = UIRefreshControl()
-    
-    private lazy var collectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        let collectionView = UICollectionView(
-            frame: .zero,
-            collectionViewLayout: layout
-        )
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        collectionView.backgroundColor = .clear
-        
-        collectionView.contentInset = UIEdgeInsets(
-            top: 16,
-            left: 16,
-            bottom: 16,
-            right: 16
-        )
-        
-        collectionView.refreshControl = refreshControl
-        collectionView.register(ShowCollectionViewCell.self)
-        
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        
-        return collectionView
-    }()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .white
-        
-        addSubviews()
-        constrainSubviews()
-        
-        
+
+    override func loadView() {
+        view = ShowsListView()
     }
 
-    func addSubviews() {
-        view.addSubview(activityIndicator)
-        view.addSubview(collectionView)
-    }
-    
-    func constrainSubviews() {
-        NSLayoutConstraint.activate(
-            [
-                activityIndicator.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-                activityIndicator.centerYAnchor.constraint(equalTo: view.centerYAnchor)
-            ]
-        )
-        
-        NSLayoutConstraint.activate(
-            [
-                collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-                collectionView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
-                collectionView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
-                collectionView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
-            ]
-        )
-        
-    }
     
 }
 
-extension ShowsListViewController: UICollectionViewDataSource {
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 20
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(
-            ShowCollectionViewCell.self,
-            for: indexPath
-        )
-        
-        return cell
-    }
-    
-    
-}
-
-extension ShowsListViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        sizeForItemAt indexPath: IndexPath
-    ) -> CGSize {
-        let width = (collectionView.bounds.width/2) - 36
-        return CGSize(
-            width: width,
-            height: 200
-        )
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return 16
-    }
-}
